+++ myus/api/db/models/base.py
@@ -1,5 +1,5 @@
 from api.db.models.comment import Comment
-from api.utils.enum.index import CommentType
+from api.utils.enum.media import CommentType
 from api.utils.functions.map import comment_type_no_map
