+++ myus/api/management/commands/sample.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from api.utils.enum.index import CommentType, CommentTypeNo
+from api.utils.enum.media import CommentType, CommentTypeNo
 from api.utils.functions.map import comment_type_map, comment_type_no_map
