+++ myus/api/src/types/schema/comment.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from pydantic import BaseModel
 from api.src.types.schema.user import AuthorOut
-from api.utils.enum.index import CommentType, CommentTypeNo
+from api.utils.enum.media import CommentType, CommentTypeNo
 
 
 class CommentListIn(BaseModel):
