+++ myus/api/utils/functions/map.py
@@ -1,4 +1,4 @@
-from api.utils.enum.index import CommentType, CommentTypeNo, MediaType
+from api.utils.enum.media import CommentType, CommentTypeNo, MediaType
 
 
 def comment_type_map(type_no: CommentTypeNo) -> CommentType:
