+++ myus/app/modules/get_form.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timedelta, timezone
 from api.db.models import AccessLog, Notification
 from api.utils.constant import model_dict
-from api.utils.enum.index import NotificationObjectType, NotificationTypeNo
+from api.utils.enum.notification import NotificationObjectType, NotificationTypeNo
 
 
 def get_client_ip(request):
